Feature/earthbeam dag  2.6

+++ edu_edfi_airflow/providers/snowflake/transfers/s3_to_snowflake.py
@@ -1,5 +1,7 @@
 from typing import Any, Optional
 
+from typing import Optional
+
 from airflow.models import BaseOperator
 from airflow.providers.snowflake.hooks.snowflake import SnowflakeHook
 from airflow.utils.decorators import apply_defaults
@@ -17,15 +19,18 @@ class S3ToSnowflakeOperator(BaseOperator):
     @apply_defaults
     def __init__(self,
         *,
-        edfi_conn_id: str,
-        snowflake_conn_id: str,
-
         tenant_code: str,
         api_year: int,
         resource: str,
         table_name: str,
 
         s3_destination_key: str,
+
+        snowflake_conn_id: str,
+
+        edfi_conn_id: Optional[str] = None,
+        ods_version: Optional[str] = None,
+        data_model_version: Optional[str] = None,
 
         xcom_return: Optional[Any] = None,
         **kwargs
@@ -42,6 +47,9 @@ class S3ToSnowflakeOperator(BaseOperator):
 
         self.s3_destination_key = s3_destination_key
 
+        self.ods_version = ods_version
+        self.data_model_version = data_model_version
+        
         self.xcom_return = xcom_return
 
 
@@ -51,20 +59,31 @@ class S3ToSnowflakeOperator(BaseOperator):
         :param context:
         :return:
         """
-        edfi_conn = EdFiHook(edfi_conn_id=self.edfi_conn_id).get_conn()
+        ### Retrieve the database and schema from the Snowflake hook.
         snowflake_hook = SnowflakeHook(snowflake_conn_id=self.snowflake_conn_id)
-
-        # Retrieve the database and schema from the Snowflake hook.
         database, schema = airflow_util.get_snowflake_params_from_conn(self.snowflake_conn_id)
 
-        if edfi_conn.is_edfi2():
-            ods_version = "ED-FI2"
-            data_model_version = "ED-FI2"
-        else:
-            ods_version = edfi_conn.get_ods_version()
-            data_model_version = edfi_conn.get_data_model_version()
+        ### Retrieve the Ed-Fi, ODS, and data model versions if not provided.
+        # (This needs to occur in execute to not call the API at every Airflow synchronize.)
+        if self.edfi_conn_id:
+            edfi_conn = EdFiHook(edfi_conn_id=self.edfi_conn_id).get_conn()
+            is_edfi2 = edfi_conn.is_edfi2()
 
-        # Build the SQL queries to be passed into `Hook.run()`.
+            if not self.ods_version:
+                self.ods_version = 'ED-FI2' if is_edfi2 else edfi_conn.get_ods_version()
+
+            if not self.data_model_version:
+                self.data_model_version = 'ED-FI2' if is_edfi2 else edfi_conn.get_data_model_version()
+
+        else:
+            is_edfi2 = False
+
+        if not (self.ods_version and self.data_model_version):
+            raise Exception(
+                f"Arguments `ods_version` and `data_model_version` could not be retrieved and must be provided."
+            )
+
+        ### Build the SQL queries to be passed into `Hook.run()`.
         qry_delete = f"""
             DELETE FROM {database}.{schema}.{self.table_name}
             WHERE tenant_code = '{self.tenant_code}'
@@ -72,20 +91,24 @@ class S3ToSnowflakeOperator(BaseOperator):
             AND name = '{self.resource}'
         """
 
+        # Brackets in regex conflict with string formatting.
+        date_regex = "\\d{8}"
+        ts_regex   = "\\d{8}T\\d{6}"
+
         qry_copy_into = f"""
             COPY INTO {database}.{schema}.{self.table_name}
                 (tenant_code, api_year, pull_date, pull_timestamp, file_row_number, filename, name, ods_version, data_model_version, v)
             FROM (
                 SELECT
-                    SPLIT_PART(metadata$filename, '/', 1) as tenant_code,
-                    SPLIT_PART(metadata$filename, '/', 2) as api_year,
-                    TO_DATE(SPLIT_PART(metadata$filename, '/', 3), 'YYYYMMDD') AS pull_date,
-                    TO_TIMESTAMP(SPLIT_PART(metadata$filename, '/', 4), 'YYYYMMDDTHH24MISS') AS pull_timestamp,
+                    '{self.tenant_code}' AS tenant_code,
+                    '{self.api_year}' AS api_year,
+                    TO_DATE(REGEXP_SUBSTR(metadata$filename, '{date_regex}'), 'YYYYMMDD') AS pull_date,
+                    TO_TIMESTAMP(REGEXP_SUBSTR(metadata$filename, '{ts_regex}'), 'YYYYMMDDTHH24MISS') AS pull_timestamp,
                     metadata$file_row_number AS file_row_number,
                     metadata$filename AS filename,
                     '{self.resource}' AS name,
-                    '{ods_version}' AS ods_version,
-                    '{data_model_version}' as data_model_version,
+                    '{self.ods_version}' AS ods_version,
+                    '{self.data_model_version}' AS data_model_version,
                     t.$1 AS v
                 FROM @{database}.util.airflow_stage/{self.s3_destination_key}
                 (file_format => 'json_default') t
@@ -93,9 +116,9 @@ class S3ToSnowflakeOperator(BaseOperator):
             force = true;
         """
 
-        # Break off prematurely if EdFi 3+ and not a full refresh.
+        ### Commit the update queries to Snowflake.
         # Incremental runs are only available in EdFi 3+.
-        if edfi_conn.is_edfi2() or airflow_util.is_full_refresh(context):
+        if is_edfi2 or airflow_util.is_full_refresh(context):
             snowflake_hook.run(
                 sql=[qry_delete, qry_copy_into],
                 autocommit=False
